Create new app Users, fix some bugs in settings file and run server.
