+++ salatik/salatik/urls.py
@@ -1,7 +1,9 @@
+from django.contrib import admin
 from django.urls import path
 from core.views import ingredient_list
 
 urlpatterns = [
-    path('/', ingredient_list, name='ingredient_list'),
+    path('admin/', admin.site.urls),
+    path('', ingredient_list, name='ingredient_list'),
     # Другие пути...
 ]
